+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.3.0 - 2023-08-29
+
 ## Added
 - Add option to save users.
 - Add option to save assignees for a merge request.